Fix braille plot y-axis scaling
Braille Mode line charts didn't draw the dots in the correct y-axis
location. For example, I created a chart with data values between 1 and
2 stepping by .1 (1, 1.1, 1.2 ... 2.0). The PlotMarkerDot mode
accurately displayed the dots in the 1 - 2 range on the Y axis. However,
when using PlotMarkerBraille mode, the dots appeared between the range of
0 and 1. The gap is much larger if the the data values are bigger (for
example, 100 to 101 stepping by .1)

I modified the braille plotting code to more closely match the termui
library implementation, which didn't have the problem.

Graphs are now displaying as expected.

Signed-off-by: Aaron DeMille <ajdemille2@gmail.com>